% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{acf}
\alias{acf}
\title{Autocovariance Function Estimation}
\usage{
acf(ts, lag.max = NULL)
}
\arguments{
\item{ts}{A numeric vector containing a time series or an object of class
"arma" ARMA(p,q) model.}

\item{lag.max}{An integer specifying the maximum lag, for which the
autocovariance is estimated}
}
\value{
A numeric vector containing the sample autocovariance function.
}
\description{
The function acf computes an estimate of the autocovariance function.
}
\details{
The autocovariance is a function that gives the covariance of the process
with itself at pairs of time points. For more detials see the long
dokumentation in the vignette "Dokumentation".
}
\examples{
acf(arma_sim(phi = c(0.5,-0.1), theta = c(0.1,0.2,-0.3), n=100))
}
