% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf_arma_ankai.R
\name{acf_arma}
\alias{acf_arma}
\title{Autocorrelation Function for Causaly ARMA(p,q) Process}
\usage{
acf_arma(phi, theta, lag = 20, dig = 5)
}
\arguments{
\item{phi}{A vector, containing coefficient of AR part as array. it require to have 1 as leading coefficient.}

\item{theta}{A vector, containing coefficient of MA part as array. it require to have 1 as leading coefficient.}

\item{lag}{The desired maximun lag , The default value is 20.}

\item{dig}{The desired rounding digit of return value, the default value is 5}
}
\value{
lag       (lag)

gamma     (autocovariance)

rho       (autocorrelation)
}
\description{
The function acf_arma() computes estimates of the autocorrelation/autocoveriance up to given lag.

It returns a data frame including lags, autocovariances and autocorellations.
}
\note{
If an integer (within 8 decimal places) is not entered for the value of dig,
  then the floor of dig will be used instead.

If an positive integer (within 8 decimal places) is not entered for the value of lag,
  then the floor of absolute value of lag will be used instead.

This function STRONGLY REQUIRE arma process to be causal.
}
\examples{
# Define time series data in a vector, x
phi <- c(1, -1, 1 / 4)
theta <- c(1, 1)

# Use acf_arma function
acf_arma(phi, theta)
acf_arma(phi, theta, 20, 8)

# Another example
phi <- c(1, rnorm(100))
theta <- c(1,rnorm(50))

acf_arma(phi, theta)
acf_arma(phi, theta, 10, 8)

}
\references{
Brockwell, P. J., & Davis, R. A. (2013). Time series: theory and methods. New York, NY: Springer.
}
\seealso{
None
}
\author{
Ankai Liu, \email{ankailiu@trentu.ca}
}
\keyword{ACF,}
\keyword{ARMA}
\keyword{Autocorrelation,}
\keyword{Autocovariance,}
