% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_acf_Melissa.R
\name{arma_acf}
\alias{arma_acf}
\title{ACF calculator for ARMA Processes}
\usage{
arma_acf(phi, theta, which_form = "acf")
}
\arguments{
\item{phi}{A numeric vector of the phi coefficients.}

\item{theta}{A numeric vector of the theta coefficients.}

\item{which_form}{Optional argument: either acv or acf; indicates which function is returned to the user.}
}
\value{
Either the acv or the acf of the ARMA process.
}
\description{
Takes a phi vector and theta vector and returns the acv or acf of the ARMA process.
}
\examples{
# Create phi and theta vectors
phi <- c(1, -1.3, 0.4)
theta <- c(1, 1)

# Use function
arma_acf(phi, theta, which_form = "acv")

}
\references{
Brockwell, P. J., & Davis, R. A. (2013). Time series: Theory and Methods. New York, NY: Springer.
}
\author{
Melissa Van Bussel <melissavanbussel@trentu.ca>
}
\keyword{ACF,}
\keyword{ACV,}
\keyword{AR,}
\keyword{ARMA,}
\keyword{Autocorrelation}
\keyword{Autocovariance,}
\keyword{MA,}
