% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_sim.R
\name{arma_sim}
\alias{arma_sim}
\title{Simulate ARMA-Process}
\usage{
arma_sim(
  phi = NULL,
  theta = NULL,
  mu = 0,
  n,
  innov.gen = rnorm,
  innov = NULL,
  burnin = NULL,
  ...
)
}
\arguments{
\item{phi, theta}{a numeric vector specifying the AR(MA)-Coefficients of an
ARMA(p,q) model.}

\item{mu}{a numeric vector specifying the mean of the ARMA(p,q)-Series.
Default is zero mean.}

\item{n}{an integer specifying the length of the resulting time series.}

\item{innov.gen}{a function from which the random innovations are drawn.}

\item{innov}{an optional time series of innovations. If not provided,
rand.gen is used}

\item{burnin}{an integer specifying the number of datapoints that are going
to be discarded, so that the characteristics of final series do not depend
on the initial values.}
}
\value{
object of class "arma" containing the simulated arma series, the
innovation series and the specified parameters.
}
\description{
Simulate from an ARMA model.
}
\details{
The ARMA model is the description of a stochastic process in terms of two
plynomials, one for the autoregression (AR) and the second for the moving
average (MA). These polynomials can be parametrized in this function, which
gives than a simulated time series with the specified characteristics. For
more detials see the long dokumentation in the vignette "Dokumentation".
}
\examples{
arma_sim(phi = c(0.5,-0.1), theta = c(0.1,0.2,-0.3), n=100)
}
