% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL.R
\name{DL}
\alias{DL}
\title{Durbin-Levinson Algorithm}
\usage{
DL(ts, p = NULL)
}
\arguments{
\item{ts}{A numeric vector containing a time series or an object of class "arma".}

\item{p}{Number of recursions or number of phi's to be determined.}
}
\value{
Numeric vector containing the phi-Coefficients determined by the Durbin-Levinson algorithm.
}
\description{
Description
}
\details{
Details
}
\examples{
DL(arma_sim(phi = c(0.8,-0.3),n = 1000,burnin = 1000))
}
