% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melissa_durbin_levinson.R
\name{durbin_levinson}
\alias{durbin_levinson}
\title{Durbin Levinson Calculator}
\usage{
durbin_levinson(gamma, X)
}
\arguments{
\item{gamma}{A numeric vector of autocovariances, beginning at gamma(0) and ending at gamma(n)}

\item{X}{A numeric vector of realizations, beginning at X_1 and ending at X_n}
}
\value{
The prediction for X_(n+1)
}
\description{
Takes a vector of autocovariances, gamma, and a vector of realizations, X, and computes a prediction for the next realization, using the Durbin Levinson algorithm.
}
\examples{
# vectors
X <- c(6, 7, 8, 9) # length of X should be one less than length of gamma
gamma <- c(5, 4.3, 3.2, 2.7, 1.6)

# Use function
durbin_levinson(gamma = gamma, X = X)

}
\references{
Brockwell, P. J., & Davis, R. A. (2013). Time series: Theory and Methods. New York, NY: Springer.
}
\author{
Melissa Van Bussel <melissavanbussel@trentu.ca>
}
\keyword{ARMA}
\keyword{Durbin}
\keyword{Durbin,}
\keyword{Levinson,}
