% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melissa_durbin_levinson_2.R
\name{durbin_levinson_2}
\alias{durbin_levinson_2}
\title{Durbin Levinson Calculator Version 2}
\usage{
durbin_levinson_2(x)
}
\arguments{
\item{x}{A numeric vector of realizations, beginning at x_1 and ending at x_N}
}
\value{
The prediction for x_N
}
\description{
Takes a vector of realizations, x, and uses it to create a vector of sample autocorrelations. The sample autocorrelations are then used to predict the next realization, via the Durbin-Levinson algorithm.
}
\note{
If the length of x is N, the length of the sample autocorrelation vector will be N-1. This means that the Durbin-Levinson algorithm is calculating a prediction for the Nth value in the time series, not the (N+1)th.
}
\examples{
# vector
x_vector <- c(6, 7, 8, 9)

# Use function
durbin_levinson_2(x = x_vector)

}
\references{
Brockwell, P. J., & Davis, R. A. (2013). Time series: Theory and Methods. New York, NY: Springer.
}
\author{
Melissa Van Bussel <melissavanbussel@trentu.ca>
}
\keyword{ARMA}
\keyword{Durbin}
\keyword{Durbin,}
\keyword{Levinson,}
