% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/innovation.R
\name{innovation}
\alias{innovation}
\title{Innovation Algorithm}
\usage{
innovation(ts, lag.max = NA)
}
\arguments{
\item{ts}{A numeric vector containing a time series or an object of class "arma".}

\item{lag.na}{Number of recursions to determined prediction.}
}
\value{
A list with the Matrix with the calculated Thetas and a vector with a mean squared error
}
\description{
Description
The innovation algorithm is algorithm, which determinate thetas and mean squared errors.They are used to predict the next values in the function innovation_prediction.
General the innovation algorithm is good for Ma-processes
}
\details{
Details
The innovation algorithm is algorithm, which determinate from previous elements and their predictors the predictor of the (n-th) next element.
Therefore the algorithm need coefficients (Thetas), which are determinate in this algorithm.
They are determinate with help from the previous Thetas, the ACF-function and mean squared errors (between predictor and the elements).
The mean square errors are also calculated in this algorithm.
}
\examples{
innovation(arma_sim(theta = c(0.8,-0.3),n = 1000,burnin = 1000))
}
