% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Innovation_prediction.R
\name{innovation_prediction}
\alias{innovation_prediction}
\title{Innovation Prediction}
\usage{
innovation_prediction(ts, steps = 1, lag.max = NA)
}
\arguments{
\item{steps}{shows which x will be predicted}

\item{lag.na}{Number of recursions to determined Prediction. (The first step's-values are needed to detect prediction)}
}
\value{
Numeric vector containing the prediction determined by the innovation algorithm.
}
\description{
Description
The innovation prediction use the innovation algorithm to predict the n-th Element after.
}
\details{
Details
The first step's-values are needed to detect prediction and the next lag.next values are used to detect the prediction.

@param ts A numeric vector containing a time series or an object of class "arma".
}
\examples{
innovation(arma_sim(theta = c(0.8,-0.3),n = 1000,burnin = 1000))
}
