\name{internals}

%\alias{.align.timeSeries}
% \alias{.appendList}
% \alias{.axTicksByTime2}
\alias{.colorwheelPalette}
% \alias{.endpoints2}
% \alias{.ohlcDailyPlot}
% \alias{.periodicity2}
% \alias{.plotOHLC}
% \alias{.xtsPlot}


\title{Exported internal functions}

\description{
  Internal functions which are still exported only because some packages
  are using them. The intent is to stop exporting or removing them.
}

\usage{
.colorwheelPalette(n)
}

\arguments{
  \item{n}{ }
}

\details{

  If you think that any of these functions are useful and don't have
  exported analogs, please contact the maintainer.
  
}
\keyword{internal}
