\name{base-methods}

\title{Methods for 'timeSeries' object}

\docType{methods}

\alias{mean,timeSeries-method}
\alias{summary,timeSeries-method}


\description{

    Methods for function in Package `base' for \code{timeSeries} object.

}

\section{Methods}{
    \describe{
    \item{x = "timeSeries"}{a \code{timeSeries} object.}
}}


\examples{
## None -
}


\keyword{chron}

