% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Periodogram.R
\name{periodogram}
\alias{periodogram}
\title{The Periodogram}
\usage{
periodogram(ts)
}
\arguments{
\item{ts}{A numeric vector or a structur of class arma which contains the
time series that should be analysed.}
}
\value{
A vector which contains the values of the Periodogram of the signal
or the time series \code{ts} at the Fourier frequencys.
}
\description{
The Periodogram estimates the power spectral density of a signal (e.g. a time
series or a series of observations). Note that the Periodogram is not a consistent estimator.
}
\examples{
n<- 1:1000
x <- rnorm(1:1000) + 2*sin(2*pi*n)
periodogram(x)
}
