% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_periodogram.R
\name{plot_periodogram}
\alias{plot_periodogram}
\title{Plot the Periodogram}
\usage{
plot_periodogram(pd)
}
\arguments{
\item{pd}{A numeric vector that contains the values of the periodogram of a
signal at the Fourier frequencies.}
}
\value{
A graph of the periodogram at the Fourier frequencys.
}
\description{
This function plots the periodogram of a signal (e.g. a time
series or a series of observations) at the Fourier frequencys between 0 and \code{pi}.
}
\examples{
n<- 1:1000
x <- rnorm(1:1000) + 2*sin(2*pi*n)
peri_x <- periodogram(x)
plot_periodogram(peri_x)
}
