% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Time Series Plot}
\usage{
plot_ts(ts)
}
\arguments{
\item{ts}{a numeric vector containing a time series or an object of class
"arma"}
}
\value{
a numeric vector containing the sample autocovariance function
}
\description{
The function plots a given Time Series.
}
\details{
Details
}
\examples{
plot_ts(arma_sim(phi = c(0.5,-0.1), theta = c(0.1,0.2,-0.3), n=100))
}
