% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtseries.R
\name{qtseries}
\alias{qtseries}
\title{Quick Time Series Plot}
\usage{
qtseries(
  data,
  time,
  voi,
  group,
  round_values = FALSE,
  verbose = FALSE,
  default_theme = TRUE
)
}
\arguments{
\item{data}{A data frame or tibble where the inputs \emph{time}, \emph{voi} and \emph{group} are stored.}

\item{time}{A numeric vector that represents time (year, month, day), stored in the data set \emph{data}.}

\item{voi}{A continuous, numeric vector stored in the data set \emph{data} (this is your variable of interest that might have interesting variations over time).}

\item{group}{A character or factor vector that is used to group the numeric data, \emph{voi}, in the specified data set, \emph{data}.}

\item{round_values}{An optional argument that rounds the variable of interest (\emph{voi}) to 3 digits (default = FALSE).}

\item{verbose}{An optional argument that that keeps the user up-to-date with the ongoing processes at each step of the graphing function (default = FALSE).}

\item{default_theme}{An optional argument that specifies a default theme: uses theme_minimal with bold plot title, legend title, and axis titles (default = TRUE).}
}
\value{
A ggplot object (a plot) depicting the temporal dynamics of some variable (\emph{voi}) grouped by a factor variable (\emph{group}) from the specified data set (\emph{data}).
}
\description{
Quick Time Series Plot
}
\examples{
qtseries(gapminder::gapminder, year, gdpPercap, country, round_values = TRUE,verbose = TRUE)

qtseries(gapminder::gapminder \%>\%
         dplyr::filter(continent == "Americas"),
         year, lifeExp, country, verbose = TRUE)
}
