% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ARMA.R
\name{sim_ARMA}
\alias{sim_ARMA}
\title{Simulate ARMA-Process}
\usage{
sim_ARMA(x, y)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
Description
}
\details{
Details
}
\examples{
add(1, 1)
add(10, 1)
}
