% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries-package.R
\docType{package}
\name{TimeSeries-package}
\alias{TimeSeries}
\alias{TimeSeries-package}
\title{TimeSeries: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\details{
I need more details!!
}
\section{MySection}{

The content of My Section :)
}

\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
