% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{timeseries}
\alias{timeseries}
\title{Timeseries}
\usage{
timeseries(data, time)
}
\arguments{
\item{data}{A phyloseq object.}

\item{time}{The name of the column containing the time variables (e.g. "Date")}
}
\description{
This function allows you plot relative abundance reads over time in a timeseries plot.
}
\keyword{timeseries}

