% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{acf}
\alias{acf}
\title{Autocovariance Function Estimation}
\usage{
acf(ts, lag.max = NA)
}
\arguments{
\item{ts}{A numeric vector containing a time series or an object of class "arma"
ARMA(p,q) model.}

\item{lag.max}{An integer specifying the maximum lag, for which the a}
}
\value{
A numeric vector containing the sample autocovariance function.
}
\description{
Description
}
\details{
Details
}
\examples{
arma_sim(phi = c(0.5,-0.1), theta = c(0.1,0.2,-0.3), n=100)
}
