% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arma_sim.R
\name{arma_sim}
\alias{arma_sim}
\title{Simulate ARMA-Process}
\usage{
arma_sim(phi = NULL, theta = NULL, n, innov.gen = rnorm, burnin = NA, ...)
}
\arguments{
\item{phi, theta}{A numeric vector specifying the AR(MA)-Coefficients of an
ARMA(p,q) model.}

\item{n}{An integer specifying the length of the resulting time series.}

\item{innov.gen}{A function from which the random innovations are drawn.}

\item{burnin}{An integer specifying the number of datapoints that are going
to be discarded, so that the characteristics of final series do not depend
on the initial values.}
}
\value{
Object of class "arma" containing the simulated arma series, the
innovation series and all the specified parameters.
}
\description{
Description
}
\details{
Details
}
\examples{
add(1, 1)
add(10, 1)
}
