% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DL.R
\name{DL}
\alias{DL}
\title{Durbin-Levinson Algorithm}
\usage{
DL(ts, p = NULL)
}
\arguments{
\item{ts}{a numeric vector containing a time series or an object of class
"arma".}

\item{p}{number of recursions or number of phi's to be determined.}
}
\value{
a numeric vector containing the phi-Coefficients determined by the
Durbin-Levinson algorithm.
}
\description{
Application of the Durbin-Levinson algorithm for forecasting time series.
}
\details{
The Durbin-Levinson Algorithm is a recursive algorithm based on the
autocorrelation function of the input time series. It computes the same
number of AR parameters as is the length of the input series. These estimated
AR parameters can be multiplied with the input series to produce a
one-step-ahead forecast. For more details see the long documentation in the
vignette "Dokumentation".
}
\examples{
DL(arma_sim(phi = c(0.8,-0.3),n = 1000,burnin = 1000))
}
