\name{base-methods}
\docType{methods}

\alias{mean,timeSeries-method}
\alias{summary,timeSeries-method}


\title{Methods for 'timeSeries' object}

\description{
  Methods for function in Package `base' for \code{timeSeries} object.

  var

  summary

  var
}

\section{Methods}{
\describe{
\item{x = "timeSeries"}{a \code{timeSeries} object.}
}}

