\name{timeSeries-deprecated}

\alias{seriesData} % removed

%\alias{.time.timeSeries}



\title{Deprecated functions in 'timeSeries' package}

%\alias{seriesPositions}
%\alias{newPositions<-}

\description{
  
  \tabular{ll}{

    \code{seriesData} extracts data slot from a 'timeSeries'.
    use \code{\link{as.matrix}} instead. \cr
    % \item{object}{
    %     [is][seriesData][seriesPositions][show][summary] -    
    %     an object of class \code{timeSeries}.
    %     }


    % \code{.time.timeSeries} \tab Extracts the time(s) from a
    % 'timeSeries', use \code{\link{time}} instead. \cr

    
    % \code{seriesPositions} \tab Extracts positions slot from a 'timeSeries', \cr
    % \code{newPositions<-} \tab Modifies positions of a 'timeSeries' object, \cr
  }

}

\keyword{internal}  
