\name{DataPart,timeSeries-method}


\title{DataPart,timeSeries-method}


\alias{getDataPart,timeSeries-method}
\alias{setDataPart,timeSeries-method}


\description{
     utilites called to implement object@.Data of \code{timeSeries} objects.
}


\examples{
## Load Microsoft Data -
   data(MSFT)
   X <- MSFT[1:10, 1:4]
   
## Get Data Part - 
   getDataPart(X)
}


\keyword{chron}

