\name{description}
\alias{description}


\title{Creates date and user information}


\description{

  Creates and returns a string containing user the datetime and the user
  name..
    
}


\usage{
description()
}


\examples{
## Show Default Description String - 
   description()
}


\keyword{programming}

